Fix API testcases to match public API typing fixes

+++ api-testcases/test-base64.c
@@ -10,7 +10,7 @@ void test_encode(duk_context *ctx) {
 	duk_push_string(ctx, "foo");
 	duk_push_int(ctx, 123);  /* dummy */
 	printf("base64 encode: %s\n", duk_base64_encode(ctx, -2));
-	printf("top after: %d\n", duk_get_top(ctx));  /* value + dummy */
+	printf("top after: %ld\n", (long) duk_get_top(ctx));  /* value + dummy */
 }
 
 void test_decode(duk_context *ctx) {
@@ -19,7 +19,7 @@ void test_decode(duk_context *ctx) {
 	duk_push_int(ctx, 321);  /* dummy */
 	duk_base64_decode(ctx, -2);  /* buffer */
 	printf("base64 decode: %s\n", duk_to_string(ctx, -2));
-	printf("top after: %d\n", duk_get_top(ctx));  /* value + dummy */
+	printf("top after: %ld\n", (long) duk_get_top(ctx));  /* value + dummy */
 	duk_set_top(ctx, 0);
 }
