+++ api-testcases/test-bug-multithread-valgrind.c
@@ -22,7 +22,7 @@ static int test_1(duk_context *ctx) {
 
 	duk_push_thread(ctx);
 	ctx2 = duk_require_context(ctx, -1);
-	printf("top: %d\n", duk_get_top(ctx));
+	printf("top: %ld\n", (long) duk_get_top(ctx));
 
 	duk_eval_string_noresult(ctx2, "aiee;");  /* ReferenceError */
 
@@ -34,7 +34,7 @@ static int test_2(duk_context *ctx) {
 
 	duk_push_thread_new_globalenv(ctx);
 	ctx2 = duk_require_context(ctx, -1);
-	printf("top: %d\n", duk_get_top(ctx));
+	printf("top: %ld\n", (long) duk_get_top(ctx));
 
 	duk_eval_string_noresult(ctx2, "zork;");  /* ReferenceError */
