+++ api-testcases/test-bug-push-buffer-maxsize.c
@@ -36,7 +36,7 @@ int test_1a(duk_context *ctx) {
 	buf = duk_push_buffer(ctx, SIZE_MAX, 0);
 	printf("ptr is non-NULL: %d\n", (buf != NULL ? 1 : 0));
 
-	printf("final top: %d\n", duk_get_top(ctx));
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
 	return 0;
 }
 
@@ -49,7 +49,7 @@ int test_1b(duk_context *ctx) {
 	buf = duk_push_buffer(ctx, SIZE_MAX - 8, 0);
 	printf("ptr is non-NULL: %d\n", (buf != NULL ? 1 : 0));
 
-	printf("final top: %d\n", duk_get_top(ctx));
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
 	return 0;
 }
 
@@ -62,7 +62,7 @@ int test_2a(duk_context *ctx) {
 	buf = duk_push_buffer(ctx, SIZE_MAX, 1);
 	printf("ptr is non-NULL: %d\n", (buf != NULL ? 1 : 0));
 
-	printf("final top: %d\n", duk_get_top(ctx));
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
 	return 0;
 }
 
@@ -76,14 +76,14 @@ int test_2b(duk_context *ctx) {
 	buf = duk_push_buffer(ctx, SIZE_MAX - 8, 1);
 	printf("ptr is non-NULL: %d\n", (buf != NULL ? 1 : 0));
 
-	printf("final top: %d\n", duk_get_top(ctx));
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
 	return 0;
 }
 
 #define  TEST(func)  do { \
 		printf("*** %s\n", #func); \
 		rc = duk_safe_call(ctx, (func), 0, 1); \
-		printf("rc=%d, result='%s'\n", rc, duk_to_string(ctx, -1)); \
+		printf("rc=%d, result='%s'\n", (int) rc, duk_to_string(ctx, -1)); \
 		duk_pop(ctx); \
 	} while (0)
